+++ examples/source/aws-backups.tf
@@ -8,6 +8,12 @@ variable "destination_vault_arn" {
   type        = string
 }
 
+variable "copy_recovery_point_role_arn" {
+  description = "Cross-account role ARN from destination account for copy recovery point lambda (module.destination.copy_recovery_point_role_arn)"
+  type        = string
+  default     = ""
+}
+
 data "aws_arn" "destination_vault_arn" {
   arn = var.destination_vault_arn
 }
@@ -183,4 +189,19 @@ module "source" {
                                         "selection_tag": "NHSE-Enable-Backup"
                                       }
 
+  backup_plan_config_parameter_store = {
+                                        "compliance_resource_types": [
+                                          "S3"
+                                        ],
+                                        "rules": [
+                                        ],
+                                        "enable": false,
+                                        "selection_tag": "NHSE-Enable-Backup"
+                                      }
+
+  # Enable copy recovery point lambda and feed destination role ARN if provided
+  lambda_copy_recovery_point_enable          = var.copy_recovery_point_role_arn != ""
+  lambda_copy_recovery_point_assume_role_arn = var.copy_recovery_point_role_arn
+
 }
+