+++ modules/aws-backup-source/lambda_parameter_store_backup.tf
@@ -1,4 +1,4 @@
-data "aws_iam_policy_document" "lambda_assume_role" {
+data "aws_iam_policy_document" "lambda_parameter_store_assume_role" {
   count = var.backup_plan_config_parameter_store.enable ? 1 : 0
   statement {
     effect = "Allow"
@@ -83,12 +83,12 @@ data "archive_file" "lambda_parameter_store_backup_zip" {
 
 resource "aws_s3_bucket" "parameter_store_backup_storage" {
   count = var.backup_plan_config_parameter_store.enable ? 1 : 0
-  bucket = "${var.name_prefix}-parameter-store-backup"
+  bucket = "${local.resource_name_prefix}-parameter-store-backup"
 
   tags = {
     Environment                                               = var.environment_name
     Application                                               = var.project_name
-    Name                                                      = "${var.name_prefix}"
+    Name                                                      = "${local.resource_name_prefix}-parameter-store-backup"
     "${var.backup_plan_config_parameter_store.selection_tag}" = var.backup_plan_config_parameter_store.selection_tag_value != null ? var.backup_plan_config_parameter_store.selection_tag_value : "True"
   }
 
@@ -107,12 +107,12 @@ resource "aws_s3_bucket_versioning" "parameter_store_backup_versioning" {
 resource "aws_iam_role" "iam_for_lambda_parameter_store_backup" {
   count = var.backup_plan_config_parameter_store.enable ? 1 : 0
   name               = "parameter_store_lambda_encryption_role"
-  assume_role_policy = data.aws_iam_policy_document.lambda_assume_role[0].json
+  assume_role_policy = data.aws_iam_policy_document.lambda_parameter_store_assume_role[0].json
 }
 
 resource "aws_iam_role_policy" "lambda_parameter_store_backup_iam_permissions" {
   count  = var.backup_plan_config_parameter_store.enable ? 1 : 0
-  name   = "${var.name_prefix}_lambda_parameter_store_backup_iam_permissions_policy"
+  name   = "${local.resource_name_prefix}_lambda_parameter_store_backup_iam_permissions_policy"
   role   = aws_iam_role.iam_for_lambda_parameter_store_backup[0].id
   policy = data.aws_iam_policy_document.lambda_parameter_store_backup_permissions[0].json
 }
@@ -121,7 +121,7 @@ resource "aws_lambda_function" "lambda_parameter_store_backup" {
   count            = var.backup_plan_config_parameter_store.enable ? 1 : 0
   filename         = data.archive_file.lambda_parameter_store_backup_zip[0].output_path
   source_code_hash = data.archive_file.lambda_parameter_store_backup_zip[0].output_base64sha256
-  function_name    = "${var.name_prefix}-parameter_store_backup"
+  function_name    = "${local.resource_name_prefix}-parameter_store_backup"
   role             = aws_iam_role.iam_for_lambda_parameter_store_backup[0].arn
   handler          = "parameter_store_backup.lambda_handler"
   runtime          = "python3.12"
@@ -136,33 +136,33 @@ resource "aws_lambda_function" "lambda_parameter_store_backup" {
   }
 }
 
-resource "aws_cloudwatch_event_rule" "aws_backup_event_rule" {
+resource "aws_cloudwatch_event_rule" "aws_backup_parameter_store_event_rule" {
   count = var.backup_plan_config_parameter_store.enable ? 1 : 0
-  name        = "${var.name_prefix}-parameter-store-backup-rule"
+  name        = "${local.resource_name_prefix}-parameter-store-backup-rule"
   description = "Triggers the Parameter Store Backup lambda."
 
   schedule_expression = "cron(${var.backup_plan_config_parameter_store.lambda_backup_cron})"
 }
 
-resource "aws_cloudwatch_event_target" "lambda_target" {
+resource "aws_cloudwatch_event_target" "lambda_parameter_store_target" {
   count     = var.backup_plan_config_parameter_store.enable ? 1 : 0
-  rule      = aws_cloudwatch_event_rule.aws_backup_event_rule[0].name
+  rule      = aws_cloudwatch_event_rule.aws_backup_parameter_store_event_rule[0].name
   arn       = aws_lambda_function.lambda_parameter_store_backup[0].arn
-  target_id = "${var.name_prefix}parameterStoreBackupLambdaTarget"
+  target_id = "${local.resource_name_prefix}parameterStoreBackupLambdaTarget"
 }
 
-resource "aws_lambda_permission" "allow_eventbridge" {
+resource "aws_lambda_permission" "lambda_parameter_store_allow_eventbridge" {
   count         = var.backup_plan_config_parameter_store.enable ? 1 : 0
-  statement_id  = "${var.name_prefix}AllowExecutionFromEventbridge"
+  statement_id  = "${local.resource_name_prefix}AllowExecutionFromEventbridge"
   action        = "lambda:InvokeFunction"
   function_name = aws_lambda_function.lambda_parameter_store_backup[0].function_name
   principal     = "events.amazonaws.com"
 
-  source_arn    = aws_cloudwatch_event_rule.aws_backup_event_rule[0].arn
+  source_arn    = aws_cloudwatch_event_rule.aws_backup_parameter_store_event_rule[0].arn
 }
 
 resource "aws_cloudwatch_log_group" "parameter_store_backup" {
   count             = var.backup_plan_config_parameter_store.enable ? 1 : 0
-  name              = "/aws/lambda/${var.name_prefix}-parameter-store-backup"
+  name              = "/aws/lambda/${local.resource_name_prefix}-parameter-store-backup"
   retention_in_days = 30
 }